+++ Composer/packages/client/src/recoilModel/dispatchers/qna.ts
@@ -94,7 +94,7 @@ export const qnaDispatcher = () => {
       const qnaFile = qnaFiles.find((f) => f.id === id);
       set(qnaAllUpViewStatusState, QnAAllUpViewStatus.Loading);
       try {
-        const response = await httpClient.get(`/helper/qna/parse`, {
+        const response = await httpClient.get(`/utilities/qna/parse`, {
           params: { url },
         });
         const content = qnaFile ? qnaFile.content + '\n' + response.data : response.data;
